third_party/blink/web_tests/fast/dom/shadow/pseudo-host-parameter-matches-shadow-host-expected.txt

Test that :host(selector) can match a shadow host when the host matches the selector: http://crbug.com/313935.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS borderColorOf(getNodeInComposedTree("host")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host")) is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
